feat(fsrs): integrate diesel for sqlite persistence
- wrap `ulid::Ulid` in a newtype with `FromSql`/`ToSql` impls
- add `FromSql`/`ToSql` impls for `CardState`
- derive `Queryable`, `Selectable`, `Insertable` for `Card`
- change `due` and `last_review` from `u64` to `i64`, stored as `bigint`
- expose `yezi_db::schema` module and depend on `yezi-db`

+++ crates/yezi-db/migrations/2026-08-15-132529-0000_cards/up.sql
@@ -2,8 +2,8 @@
 create table cards (
     id blob primary key not null,
     state integer not null,
-    due integer not null,
-    last_review integer not null,
+    due bigint not null,
+    last_review bigint not null,
     stability real not null,
     difficulty real not null
 );
